Maven/metadata rule unclaimed maintainer email domain on local and remote projects (#26)
Co-authored-by: Garance <garance@MacBook-Air-de-Garance.local>

Move metadata rule unclaimed maintainer's email domain into sourcecode rules

Since metadata rules are run only on remote packages, it is relevant that this one is run on local and remote packages as the maintainer's email can be found in the pom, which is researched locally. It gives more preciseness also on local packages.